Wow actual robot control
